feat: blogs logic and fix bugs

+++ src/modules/blogs/service/blogs.service.ts
@@ -1,5 +1,5 @@
 import {
-  BadRequestException,
+  ForbiddenException,
   Injectable,
   NotFoundException,
 } from '@nestjs/common';
@@ -7,13 +7,14 @@ import { InjectModel } from '@nestjs/mongoose';
 import { Blogs, BlogsDocument, BlogsViewModel } from '../schemas/blogs.schema';
 import { BlogsRepository } from '../repository/blogs.repository';
 import { BlogPaginationDto } from '../../helpers/dto/pagination.dto';
-import { BlogsDto, DB_BlogsType } from '../dto/blogsDto';
+import { BlogsDto, BlogsModal_For_DB } from '../dto/blogsDto';
 import { Model } from 'mongoose';
 import { ObjectId } from 'mongodb';
 import { PaginationViewModel } from '../../helpers/pagination/pagination-view-model';
 import { PostsViewModal } from '../../posts/schemas/posts.schema';
 import { CreatePostDto } from '../../posts/dto/createPostDto';
 import { PostsRepository } from '../../posts/repository/posts.repository';
+import { UserModel } from '../../users/schemas/users.schema';
 
 @Injectable()
 export class BlogsService {
@@ -29,22 +30,28 @@ export class BlogsService {
     return this.blogsRepository.getBlogs(paginationType);
   }
 
-  async createBlog(blog: BlogsDto): Promise<BlogsViewModel> {
-    const newBlog: DB_BlogsType = {
+  async createBlog(blog: BlogsDto, user: UserModel): Promise<BlogsViewModel> {
+    const newBlog: BlogsModal_For_DB = {
       id: new ObjectId().toString(),
       name: blog.name,
       description: blog.description,
       websiteUrl: blog.websiteUrl,
       createdAt: new Date().toISOString(),
       isMembership: false,
+      blogOwnerInfo: {
+        userId: user.id,
+        userLogin: user.login,
+      },
     };
-    return this.blogsRepository.createBlog(newBlog);
+    return this.blogsRepository.createBlog({ ...newBlog });
   }
 
-  async deleteBlogById(id: string): Promise<boolean> {
+  async deleteBlogById(id: string, userId: string): Promise<boolean> {
     const findBlogById = await this.blogsRepository.findBlogById(id);
     if (!findBlogById)
       throw new NotFoundException(`User with ID ${id} not found`);
+    if (findBlogById.blogOwnerInfo.userId !== userId)
+      throw new ForbiddenException([]);
     return this.blogsRepository.deleteBlogById(id);
   }
 
@@ -54,11 +61,13 @@ export class BlogsService {
       throw new NotFoundException(`User with ID ${id} not found`);
     return findBlogById;
   }
+
   async updateBlogById(id: string, user: BlogsDto): Promise<boolean> {
     const result = await this.blogsRepository.updateBlogById(id, user);
     if (!result) throw new NotFoundException(`User with ID ${id} not found`);
     return result;
   }
+
   async createPostByBlogId(
     blogId: string,
     newPostByBlogId: CreatePostDto,
